+++ frontend/src/components/authz/PermissionRequired/index.tsx
