+++ frontend/src/components/authz/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginRequired } from "./LoginRequired";
+export { default as PermissionRequired } from "./PermissionRequired";
